Multiple modules done as per 31012019

+++ controllers/CauseListController.php
@@ -7,7 +7,10 @@ use app\models\CauseList;
 use app\models\CauseListSearch;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
+use yii\web\ForbiddenHttpException;
 use yii\filters\VerbFilter;
+use yii\filters\AccessControl;
+use yii\web\UploadedFile;
 
 /**
  * CauseListController implements the CRUD actions for CauseList model.
@@ -20,6 +23,17 @@ class CauseListController extends Controller
     public function behaviors()
     {
         return [
+            'access' => [
+                'class' => AccessControl::className(),
+                'only' => ['index', 'view', 'create', 'update', 'delete', 'uploadexcel'],
+                'rules' => [
+                    [
+                        //'actions' => ['logout'],
+                        'allow' => true,
+                        'roles' => ['@'],
+                    ],
+                ],
+            ],
             'verbs' => [
                 'class' => VerbFilter::className(),
                 'actions' => [
@@ -35,13 +49,18 @@ class CauseListController extends Controller
      */
     public function actionIndex()
     {
-        $searchModel = new CauseListSearch();
-        $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
+        if(Yii::$app->user->can('IO-MR'))
+            {
+            $searchModel = new CauseListSearch();
+            $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
 
-        return $this->render('index', [
-            'searchModel' => $searchModel,
-            'dataProvider' => $dataProvider,
-        ]);
+            return $this->render('index', [
+                'searchModel' => $searchModel,
+                'dataProvider' => $dataProvider,
+            ]);
+        } else {
+            throw new ForbiddenHttpException(Yii::t('yii', "Sorry, you currently don't have privilege to this action."));
+        }       
     }
 
     /**
@@ -97,9 +116,14 @@ class CauseListController extends Controller
      */
     public function actionView($id)
     {
-        return $this->render('view', [
-            'model' => $this->findModel($id),
-        ]);
+        if(Yii::$app->user->can('IO-MR'))
+            {
+            return $this->render('view', [
+                'model' => $this->findModel($id),
+            ]);
+        } else {
+            throw new ForbiddenHttpException(Yii::t('yii', "Sorry, you currently don't have privilege to this action."));
+        }      
     }
 
     /**
@@ -109,15 +133,90 @@ class CauseListController extends Controller
      */
     public function actionCreate()
     {
-        $model = new CauseList();
+        if(Yii::$app->user->can('IO-MR'))
+            {
+            $model = new CauseList();
 
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->id]);
-        }
+            if ($model->load(Yii::$app->request->post()) && $model->save()) {
+                return $this->redirect(['view', 'id' => $model->id]);
+            }
 
-        return $this->render('create', [
-            'model' => $model,
-        ]);
+            return $this->render('create', [
+                'model' => $model,
+            ]);
+         } else {
+            throw new ForbiddenHttpException(Yii::t('yii', "Sorry, you currently don't have privilege to this action."));
+        }       
+    }
+
+    /**
+    * Here we are going to define the aaction that will be uploading the excel
+    * file to the database
+    **/
+    public function actionUploadexcel()
+    {
+        if(Yii::$app->user->can('IO-MR'))
+            {
+                $model = new CauseList();
+
+                if ($model->load(Yii::$app->request->post())) {
+
+                       //Get instance of Uploaded file
+                        $model = new CauseList();
+                        $model->excel = UploadedFile::getInstance($model, 'excel');
+                        if ($model->excel != null) {
+                                $model->excel->saveAs( 'main_registry/cause_lists/' . str_replace(' ', '_', $model->excel->baseName) . '.' . $model->excel->extension );
+                                $inputFile = 'main_registry/cause_lists/' . str_replace(' ', '_', $model->excel->baseName) . '.' . $model->excel->extension;
+                            try{
+                                $inputFileType = \PHPExcel_IOFactory::identify($inputFile);
+                                $objReader = \PHPExcel_IOFactory::createReader($inputFileType);
+                                $objPHPExcel = $objReader->load($inputFile);
+                            } catch (Exception $e) {
+                                die('Error');
+                            }
+
+                            $sheet = $objPHPExcel->getSheet(0);
+                            $highestRow = $sheet->getHighestRow();
+                            $highestColumn = $sheet->getHighestColumn();
+
+                            for($row=1; $row <= $highestRow; $row++)
+                            {
+                                $rowData = $sheet->rangeToArray('A'.$row.':'.$highestColumn.$row,NULL,TRUE,FALSE);
+
+                                if($row==1)
+                                {
+                                    continue;
+                                }
+
+                                $causelists = new CauseList();
+                                $causelists->dates = $rowData[0][0]; 
+                                $causelists->case_number  = $rowData[0][1]; 
+                                $causelists->parties  = $rowData[0][2]; 
+                                $causelists->witness  = $rowData[0][3]; 
+                                $causelists->advocate_plaintiff  = $rowData[0][4]; 
+                                $causelists->advocate_defendant  = $rowData[0][5]; 
+                                $causelists->division  = $rowData[0][6]; 
+                                $causelists->save();
+
+                                //print_r($causelists->getErrors());
+                            }
+                           // die('okay');
+
+
+                            return $this->redirect(['cause-list/index']);
+                            }else {
+                                        return $this->render('uploadexcel', [
+                                        'model' => $model,
+                                          ]);
+                                   }   
+                }
+
+                return $this->renderAjax('uploadexcel', [
+                    'model' => $model,
+                ]);
+        } else {
+            throw new ForbiddenHttpException(Yii::t('yii', "Sorry, you currently don't have privilege to this action."));
+        }           
     }
 
     /**
@@ -130,14 +229,19 @@ class CauseListController extends Controller
     public function actionUpdate($id)
     {
         $model = $this->findModel($id);
+        if(Yii::$app->user->can('IO-MR'))
+            {
 
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->id]);
-        }
+            if ($model->load(Yii::$app->request->post()) && $model->save()) {
+                return $this->redirect(['view', 'id' => $model->id]);
+            }
 
-        return $this->render('update', [
-            'model' => $model,
-        ]);
+            return $this->render('update', [
+                'model' => $model,
+            ]);
+        } else {
+            throw new ForbiddenHttpException(Yii::t('yii', "Sorry, you currently don't have privilege to this action."));
+        }     
     }
 
     /**
@@ -149,9 +253,15 @@ class CauseListController extends Controller
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
+        $model = $this->findModel($id);
+        if(Yii::$app->user->can('IO-MR'))
+            {
+                $this->findModel($id)->delete();
 
-        return $this->redirect(['index']);
+                return $this->redirect(['index']);
+        } else {
+            throw new ForbiddenHttpException(Yii::t('yii', "Sorry, you currently don't have privilege to this action."));
+        }     
     }
 
     /**
