+++ controllers/CommentsController.php
@@ -8,7 +8,9 @@ use app\models\Posts;
 use app\models\CommentsSearch;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
+use yii\web\ForbiddenHttpException;
 use yii\filters\VerbFilter;
+use yii\filters\AccessControl;
 
 /**
  * CommentsController implements the CRUD actions for Comments model.
@@ -21,6 +23,17 @@ class CommentsController extends Controller
     public function behaviors()
     {
         return [
+            'access' => [
+                'class' => AccessControl::className(),
+                'only' => ['index', 'view', 'update', 'delete'],
+                'rules' => [
+                    [
+                        //'actions' => ['logout'],
+                        'allow' => true,
+                        'roles' => ['@'],
+                    ],
+                ],
+            ],
             'verbs' => [
                 'class' => VerbFilter::className(),
                 'actions' => [
@@ -36,13 +49,38 @@ class CommentsController extends Controller
      */
     public function actionIndex()
     {
-        $searchModel = new CommentsSearch();
-        $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
+        if(Yii::$app->user->can('IO-MR'))
+            {
+            $searchModel = new CommentsSearch();
+            $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
+            $dataProvider->setSort([
+                            'attributes' => [
+                                'id' => [
+                                    'asc' => ['id' => SORT_ASC],
+                                    'desc' => ['id' => SORT_DESC],
+                                    'default' => SORT_DESC,
+                                ],
+                                
+                                'time' => [
+                                    'asc' => ['id' => SORT_ASC],
+                                    'desc' => ['id' => SORT_DESC],
+                                    'default' => SORT_DESC
+                                ],
+                                
+                            ],
+                            'defaultOrder' => [
+                                'id' => SORT_DESC,
+                                'time' => SORT_DESC,
+                            ]
+                        ]);
 
-        return $this->render('index', [
-            'searchModel' => $searchModel,
-            'dataProvider' => $dataProvider,
-        ]);
+            return $this->render('index', [
+                'searchModel' => $searchModel,
+                'dataProvider' => $dataProvider,
+            ]);
+        } else {
+            throw new ForbiddenHttpException(Yii::t('yii', "Sorry, you currently don't have privilege to this action."));
+        }     
     }
 
     /**
@@ -53,9 +91,14 @@ class CommentsController extends Controller
      */
     public function actionView($id)
     {
-        return $this->render('view', [
-            'model' => $this->findModel($id),
-        ]);
+        if(Yii::$app->user->can('IO-MR'))
+            {
+            return $this->render('view', [
+                'model' => $this->findModel($id),
+            ]);
+        } else {
+            throw new ForbiddenHttpException(Yii::t('yii', "Sorry, you currently don't have privilege to this action."));
+        }      
     }
 
     /**
@@ -65,17 +108,20 @@ class CommentsController extends Controller
      */
     public function actionCreate($id)
     {
-        $model = new Comments();
+            $model = new Comments();
+            date_default_timezone_set("Africa/Dar_es_Salaam");
 
-        if ($model->load(Yii::$app->request->post())) {
-            $model->post_id = $id;
-            $model->save();
-            return $this->redirect(['posts/webview', 'id' => $model->post_id]);
-        }
+            if ($model->load(Yii::$app->request->post())) {
+                $model->post_id = $id;
+                $model->status = 1;
+                $model->time = date('Y-m-d H:i:s');
+                $model->save();
+                return $this->redirect(['posts/webview', 'id' => $model->post_id]);
+            }
 
-        return $this->render('create', [
-            'model' => $model,
-        ]);
+            return $this->render('create', [
+                'model' => $model,
+            ]);      
     }
 
     /**
@@ -89,13 +135,41 @@ class CommentsController extends Controller
     {
         $model = $this->findModel($id);
 
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->id]);
-        }
+       if(Yii::$app->user->can('IO-MR'))
+        {
+            if ($model->load(Yii::$app->request->post()) && $model->save()) {
+                return $this->redirect(['view', 'id' => $model->id]);
+            }
 
-        return $this->render('update', [
-            'model' => $model,
-        ]);
+            return $this->render('update', [
+                'model' => $model,
+            ]);
+        } else {
+            throw new ForbiddenHttpException(Yii::t('yii', "Sorry, you currently don't have privilege to this action."));
+        }     
+    }
+
+     /**
+     * This action is for approving the new files created by users that have no Head privileges.
+     */
+    
+    public function actionApprove($id)
+    {
+        $model = $this->findModel($id);
+        
+        if(Yii::$app->user->can('IO-MR')) 
+     {   
+            
+            Yii::$app->db->createCommand()
+                    ->update('comments', 
+                            ['status' => 2,],['id' =>$id]
+                    )->execute();
+            return $this->redirect(['index']);
+            
+      } else {
+            throw new ForbiddenHttpException(Yii::t('yii', "Sorry, you currently don't have privilege to this action."));
+        }        
+    
     }
 
     /**
@@ -107,9 +181,15 @@ class CommentsController extends Controller
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
+         $model = $this->findModel($id);
+        if(Yii::$app->user->can('IO-MR'))
+        {
+            $this->findModel($id)->delete();
 
-        return $this->redirect(['index']);
+            return $this->redirect(['index']);
+        } else {
+            throw new ForbiddenHttpException(Yii::t('yii', "Sorry, you currently don't have privilege to this action."));
+        }     
     }
 
     /**
